refactor(ui): extract ModelLoadCoordinator from ChatViewModel (phase 3 of #329)
Moves the "latest-wins with cancellation" model-load state machine out of
ChatViewModel into a dedicated ModelLoadCoordinator class. ChatViewModel
becomes a thin facade: dispatchSelectedLoad, loadSelectedModel,
loadCloudEndpoint, and unloadModel all forward to the coordinator.

The five stored properties (coordinatedLoadTask, latestLoadIntentGeneration,
progressBridgePollInterval, progressBridgeMinTransitionInterval,
lastProgressTransitionInstant) now live on the coordinator. Forwarding
computed vars for progressBridgePollInterval and
progressBridgeMinTransitionInterval keep existing test code unchanged.

Closes #354
